+++ App.js
@@ -1,21 +1,9 @@
 import React from 'react';
-import { NavigationContainer } from '@react-navigation/native';
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-
-import HomeScreen from './app/screens/Home/Index';
-import SettingsScreen from './app/screens/Settings/Index';
-
-const Tab = createBottomTabNavigator();
+import RootNavigator from './app/navigation/RootNavigator';
+import './app/constants/IMLocalize';
 
 const App: () => Node = () => {
-  return (
-    <NavigationContainer>
-        <Tab.Navigator>
-            <Tab.Screen name="Home" component={HomeScreen} />
-            <Tab.Screen name="Settings" component={SettingsScreen} />
-        </Tab.Navigator>
-    </NavigationContainer>
-);
+  return <RootNavigator />;
 };
 
 export default App;