+++ app/constants/translations/de/common.js
@@ -1,0 +1,4 @@
+export default {
+  hello: 'Hallo',
+  languageSelector: 'Wähle deine Sprache',
+};