+++ app/navigation/RootNavigator.js
@@ -1,0 +1,47 @@
+import * as React from 'react';
+import {NavigationContainer} from '@react-navigation/native';
+import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
+import Ionicons from 'react-native-vector-icons/dist/Ionicons';
+import {useTranslation} from 'react-i18next';
+
+import HomeScreen from '../screens/Home/Index';
+import SettingsScreen from '../screens/Settings/Index';
+
+const Tab = createBottomTabNavigator();
+
+export default function RootNavigator() {
+  const {t} = useTranslation();
+
+  return (
+    <NavigationContainer>
+      <Tab.Navigator
+        screenOptions={({route}) => ({
+          tabBarIcon: ({focused, color, size}) => {
+            let iconName;
+
+            if (route.name === 'Home') {
+              iconName = focused ? 'ios-home' : 'ios-home-outline';
+            } else if (route.name === 'Settings') {
+              iconName = focused ? 'ios-settings' : 'ios-settings-outline';
+            }
+
+            return <Ionicons name={iconName} size={size} color={color} />;
+          },
+          tabBarActiveTintColor: 'tomato',
+          tabBarInactiveTintColor: 'gray',
+          headerShown: false,
+        })}>
+        <Tab.Screen
+          name="Home"
+          component={HomeScreen}
+          options={{tabBarLabel: t('navigate:home')}}
+        />
+        <Tab.Screen
+          name="Settings"
+          component={SettingsScreen}
+          options={{tabBarLabel: t('navigate:settings')}}
+        />
+      </Tab.Navigator>
+    </NavigationContainer>
+  );
+}