+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - [feat] First commit. The app in react native.
 - [feat] Add home screen.
 - [feat] Add settings screen.
+- [feat] Add languages in settings screen.